Remove unneccessary subfolders
