+++ front/app/containers/Admin/pagesAndMenu/containers/NavigationSettings/HiddenNavbarItemList/Header.tsx
@@ -6,7 +6,7 @@ import { Box } from '@citizenlab/cl2-component-library';
 
 // i18n
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from '../../messages';
+import messages from './messages';
 
 interface Props {
   itemsNotInNavbarPresent: boolean;
