+++ front/app/containers/Admin/pagesAndMenu/containers/NavigationSettings/HiddenNavbarItemList/HiddenNavbarItemList.test.tsx
