+++ front/app/containers/Admin/pagesAndMenu/containers/NavigationSettings/HiddenNavbarItemList/messages.ts
@@ -1,0 +1,13 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  deletePageConfirmationHidden: {
+    id: 'app.containers.Admin.PagesAndMenu.components.deletePageConfirmationHidden',
+    defaultMessage:
+      'Are you sure you want to delete this page? This cannot be undone.',
+  },
+  hiddenFromNavigation: {
+    id: 'app.containers.Admin.PagesAndMenu.components.hiddenFromNavigation',
+    defaultMessage: 'Other available pages',
+  },
+});
