--- front/app/containers/Admin/pagesAndMenu/containers/PagesMenu/NavigationSettings/HiddenNavbarItemList/messages.ts
@@ -1,9 +1,0 @@
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  deletePageConfirmationHidden: {
-    id: 'app.modules.navbar.admin.containers.deletePageConfirmationHidden',
-    defaultMessage:
-      'Are you sure you want to delete this page? This cannot be undone.',
-  },
-});
