+++ front/app/containers/Admin/pagesAndMenu/routes.tsx
@@ -5,7 +5,7 @@ import { Navigate } from 'react-router-dom';
 const CustomPagesIndex = lazy(() => import('./containers/CustomPages'));
 const PagesAndMenuIndex = lazy(() => import('containers/Admin/pagesAndMenu'));
 const NavigationSettings = lazy(
-  () => import('./containers/PagesMenu/NavigationSettings')
+  () => import('./containers/NavigationSettings')
 );
 const EditNavbarItemForm = lazy(
   () => import('./containers/EditNavbarItemForm')
